+++ src/config-utils.ts
@@ -7,6 +7,7 @@ import { coerce, satisfies } from "semver";
 import { logError, logInfo } from "./log-utils.js";
 
 export interface Config<T> {
+  source: string;
   raw: string;
   json: T;
 }
@@ -66,6 +67,7 @@ export async function getConfig<T>(cwd: string, file: string): Promise<Config<T>
     switch (fileType) {
       case ".json": {
         config = {
+          source: file,
           raw,
           json: parse(raw) as T,
         };
@@ -75,6 +77,7 @@ export async function getConfig<T>(cwd: string, file: string): Promise<Config<T>
       case ".yml": {
         const json = load(raw) as T;
         config = {
+          source: file,
           raw: JSON.stringify(json),
           json,
         };
@@ -85,6 +88,7 @@ export async function getConfig<T>(cwd: string, file: string): Promise<Config<T>
         const moduleImport = await import(filePath) as { default: T; };
         const json = moduleImport.default;
         config = {
+          source: file,
           raw: JSON.stringify(json),
           json,
         };